requirements.txt
--only-binary :all:
streamlit==1.29.0
pandas==2.1.3
numpy==1.26.0
matplotlib==3.8.2
seaborn==0.13.2
scikit-learn==1.3.2
joblib==1.3.2
pillow==10.2.0